+++ plugins/module_utils/crud_manage_vol.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 
 """
-This module performs the Unmanage operations on VMs
+This module performs the Manage operations on Volumes
 """
 
 import requests
@@ -30,27 +30,24 @@ def get_endpoint_url_by_service_name(mod, connectn, service_name, tenant_id):
         mod.fail_json(msg=f"No service found with the name '{service_name}'", changed=False)
 
 
-def unmanage_vm(mod, endpoint, vmurl, authtoken, post_data):
+def manage_vol(mod, endpoint, manage_url, authtoken, post_data):
     """
-    Performs UnManage operations on the VM provided
+    Performs Manage operations on the Vol provided
     """
     headers_scg = {"X-Auth-Token": authtoken, "Content-Type": "application/json"}
-    responce = requests.get(vmurl, headers=headers_scg, verify=False)
-    host_value = responce.json()['server']['OS-EXT-SRV-ATTR:host']
-    unmanage_url = f"{endpoint}/os-hosts/{host_value}/unmanage"
-    responce = requests.post(unmanage_url, headers=headers_scg, json=post_data, verify=False)
+    responce = requests.post(manage_url, headers=headers_scg, json=post_data, verify=False)
     if responce.ok:
-        return "VM Unmanage action is done"
+        return f"Data Volume: {post_data} Manage action is done"
     else:
         mod.fail_json(
             msg=f"An unexpected error occurred: {responce.json()}", changed=False
         )
 
 
-def unmanage_ops(mod, connectn, authtoken, tenant_id, vm_id):
-    service_name = "compute"
+def manage_ops(mod, connectn, authtoken, tenant_id, vol_id, host):
+    service_name = "volume"
     endpoint = get_endpoint_url_by_service_name(mod, connectn, service_name, tenant_id)
-    unmanage_data = {"servers": [vm_id]}
-    url = f"{endpoint}/servers/{vm_id}"
-    result = unmanage_vm(mod, endpoint, url, authtoken, unmanage_data)
+    manage_data = {"volumes": [vol_id]}
+    url = f"{endpoint}/os-hosts/{host}/onboard"
+    result = manage_vol(mod, endpoint, url, authtoken, manage_data)
     return result