+++ plugins/module_utils/crud_volume_type_update.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python
 
 """
-This module performs the Unmanage operations on VMs
+This module performs the PowerVC Volume Type Update operations
 """
 
 import requests
@@ -30,27 +30,20 @@ def get_endpoint_url_by_service_name(mod, connectn, service_name, tenant_id):
         mod.fail_json(msg=f"No service found with the name '{service_name}'", changed=False)
 
 
-def unmanage_vm(mod, endpoint, vmurl, authtoken, post_data):
+def post_volume_type(headers_vm, url, extra_specs):
     """
-    Performs UnManage operations on the VM provided
+    Performs Update Volume Type operation
     """
-    headers_scg = {"X-Auth-Token": authtoken, "Content-Type": "application/json"}
-    responce = requests.get(vmurl, headers=headers_scg, verify=False)
-    host_value = responce.json()['server']['OS-EXT-SRV-ATTR:host']
-    unmanage_url = f"{endpoint}/os-hosts/{host_value}/unmanage"
-    responce = requests.post(unmanage_url, headers=headers_scg, json=post_data, verify=False)
+    data = {'extra_specs': extra_specs}
+    responce = requests.post(url, headers=headers_vm, json=data, verify=False)
     if responce.ok:
-        return "VM Unmanage action is done"
-    else:
-        mod.fail_json(
-            msg=f"An unexpected error occurred: {responce.json()}", changed=False
-        )
+        return (f"Volume Type changes are done - {responce.json()}")
 
 
-def unmanage_ops(mod, connectn, authtoken, tenant_id, vm_id):
-    service_name = "compute"
+def volume_ops(mod, connectn, authtoken, tenant_id, vol_type_id, extra_specs):
+    service_name = "volume"
     endpoint = get_endpoint_url_by_service_name(mod, connectn, service_name, tenant_id)
-    unmanage_data = {"servers": [vm_id]}
-    url = f"{endpoint}/servers/{vm_id}"
-    result = unmanage_vm(mod, endpoint, url, authtoken, unmanage_data)
+    headers_vm = get_headers(authtoken)
+    volume_type_url = f"{endpoint}/types/{vol_type_id}/extra_specs"
+    result = post_volume_type(mod, headers_vm, volume_type_url, extra_specs)
     return result