+++ plugins/modules/clone_vm.py
@@ -75,7 +75,7 @@ EXAMPLES = '''
        - name:  Perform VM Clone Operation on VM with network and IP details
          ibm.powervc.clone_vm:
             auth: "{{ auth }}"
-            vm_name: "VM_NAME"
+            name: "VM_NAME"
             clonevm_name: "CLONEVM_NAME"
             nics:
                 - net-name: "NET-NAME"