+++ docs/source/release_notes.rst
@@ -8,8 +8,7 @@ Releases
 Version 1.0.0
 -------------
 Notes
-IBM PowerVC ansible modules included as part of version 1.0.0.
-
+* IBM PowerVC ansible modules included as part of version 1.0.0.
 - capture_vm
 - clone_vm
 - manage_vm
@@ -23,3 +22,17 @@ IBM PowerVC ansible modules included as part of version 1.0.0.
 - unmanage_vm
 - volume_attach
 - volume_detach
+
+Version 1.1.0
+-------------
+Notes
+* Enabled the option to support SRIOV in server module.
+* Below Storage modules are added as part of this version 1.1.0.
+- volume_type_info
+- volume_type_update
+- copy_volume_type
+- all_volumes
+- manage_volume
+- unmanage_volume
+- snapshot_info 
+